requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
django-cors-headers
django-environ
markdown
django-filter
django-guardian
django-debug-toolbar
django-extensions
djangorestframework-gis
requests